+++ sdk/core/azure-core/test/ut/context.cpp
@@ -25,7 +25,7 @@ TEST(Context, BasicBool)
   Context context;
   // New context from previous
   auto c2 = context.WithValue("key", true);
-  auto& value = c2.Get<bool>("key");
+  auto& value = c2.GetValue<bool>("key");
   EXPECT_TRUE(value == true);
 }
 
@@ -34,7 +34,7 @@ TEST(Context, BasicInt)
   Context context;
   // New context from previous
   auto c2 = context.WithValue("key", 123);
-  auto& value = c2.Get<int>("key");
+  auto& value = c2.GetValue<int>("key");
   EXPECT_TRUE(value == 123);
 }
 
@@ -45,7 +45,7 @@ TEST(Context, BasicStdString)
   Context context;
   // New context from previous
   auto c2 = context.WithValue("key", s);
-  auto& value = c2.Get<std::string>("key");
+  auto& value = c2.GetValue<std::string>("key");
   EXPECT_TRUE(value == s);
 }
 
@@ -57,7 +57,7 @@ TEST(Context, BasicChar)
   Context context;
   // New context from previous
   auto c2 = context.WithValue("key", str);
-  auto& value = c2.Get<const char*>("key");
+  auto& value = c2.GetValue<const char*>("key");
   EXPECT_TRUE(value == s);
   EXPECT_TRUE(value == str);
 }
@@ -141,13 +141,13 @@ TEST(Context, Chain)
   auto c7 = c6.WithValue("c7", "7");
   auto finalContext = c7.WithValue("finalContext", "Final");
 
-  auto& valueT2 = finalContext.Get<int>("c2");
-  auto& valueT3 = finalContext.Get<int>("c3");
-  auto& valueT4 = finalContext.Get<int>("c4");
-  auto& valueT5 = finalContext.Get<const char*>("c5");
-  auto& valueT6 = finalContext.Get<const char*>("c6");
-  auto& valueT7 = finalContext.Get<const char*>("c7");
-  auto& valueT8 = finalContext.Get<const char*>("finalContext");
+  auto& valueT2 = finalContext.GetValue<int>("c2");
+  auto& valueT3 = finalContext.GetValue<int>("c3");
+  auto& valueT4 = finalContext.GetValue<int>("c4");
+  auto& valueT5 = finalContext.GetValue<const char*>("c5");
+  auto& valueT6 = finalContext.GetValue<const char*>("c6");
+  auto& valueT7 = finalContext.GetValue<const char*>("c7");
+  auto& valueT8 = finalContext.GetValue<const char*>("finalContext");
 
   EXPECT_TRUE(valueT2 == 123);
   EXPECT_TRUE(valueT3 == 456);
@@ -165,8 +165,8 @@ TEST(Context, MatchingKeys)
   auto c2 = context.WithValue("key", 123);
   auto c3 = c2.WithValue("key", 456);
 
-  auto& valueT2 = c2.Get<int>("key");
-  auto& valueT3 = c3.Get<int>("key");
+  auto& valueT2 = c2.GetValue<int>("key");
+  auto& valueT3 = c3.GetValue<int>("key");
 
   EXPECT_TRUE(valueT2 == 123);
   EXPECT_TRUE(valueT3 == 456);
@@ -180,7 +180,7 @@ struct SomeStructForContext
 TEST(Context, InstanceValue)
 {
   auto contextP = Context::GetApplicationContext().WithValue("struct", SomeStructForContext());
-  auto& contextValueRef = contextP.Get<SomeStructForContext>("struct");
+  auto& contextValueRef = contextP.GetValue<SomeStructForContext>("struct");
   EXPECT_EQ(contextValueRef.someField, 12345);
 }
 
@@ -188,7 +188,7 @@ TEST(Context, UniquePtr)
 {
   auto contextP = Context::GetApplicationContext().WithValue(
       "struct", std::make_unique<SomeStructForContext>());
-  auto& contextValueRef = contextP.Get<std::unique_ptr<SomeStructForContext>>("struct");
+  auto& contextValueRef = contextP.GetValue<std::unique_ptr<SomeStructForContext>>("struct");
   EXPECT_EQ(contextValueRef->someField, 12345);
 }
 
@@ -202,31 +202,31 @@ TEST(Context, HeapLinkIntegrity)
 
     auto secondGeneration = firstGeneration.WithValue("b", std::string("b"));
     EXPECT_TRUE(secondGeneration.HasKey("a"));
-    EXPECT_EQ("a", secondGeneration.Get<std::string>("a"));
+    EXPECT_EQ("a", secondGeneration.GetValue<std::string>("a"));
     EXPECT_TRUE(secondGeneration.HasKey("b"));
-    EXPECT_EQ("b", secondGeneration.Get<std::string>("b"));
+    EXPECT_EQ("b", secondGeneration.GetValue<std::string>("b"));
 
     // Now overide the generation
     secondGeneration = secondGeneration.WithValue("c", std::string("c"));
     EXPECT_TRUE(
         secondGeneration.HasKey("a")); // Still know about first gen - The link is still in heap
-    EXPECT_EQ("a", secondGeneration.Get<std::string>("a"));
+    EXPECT_EQ("a", secondGeneration.GetValue<std::string>("a"));
     EXPECT_TRUE(secondGeneration.HasKey(
         "b")); // Still knows about the initial second gen, as a shared_ptr, it is still on heap
-    EXPECT_EQ("b", secondGeneration.Get<std::string>("b"));
+    EXPECT_EQ("b", secondGeneration.GetValue<std::string>("b"));
     EXPECT_TRUE(secondGeneration.HasKey("c")); // Check new value
-    EXPECT_EQ("c", secondGeneration.Get<std::string>("c"));
+    EXPECT_EQ("c", secondGeneration.GetValue<std::string>("c"));
 
     // One more override
     secondGeneration = secondGeneration.WithValue("d", std::string("d"));
     EXPECT_TRUE(secondGeneration.HasKey("a"));
-    EXPECT_EQ("a", secondGeneration.Get<std::string>("a"));
+    EXPECT_EQ("a", secondGeneration.GetValue<std::string>("a"));
     EXPECT_TRUE(secondGeneration.HasKey("b"));
-    EXPECT_EQ("b", secondGeneration.Get<std::string>("b"));
+    EXPECT_EQ("b", secondGeneration.GetValue<std::string>("b"));
     EXPECT_TRUE(secondGeneration.HasKey("c"));
-    EXPECT_EQ("c", secondGeneration.Get<std::string>("c"));
+    EXPECT_EQ("c", secondGeneration.GetValue<std::string>("c"));
     EXPECT_TRUE(secondGeneration.HasKey("d"));
-    EXPECT_EQ("d", secondGeneration.Get<std::string>("d"));
+    EXPECT_EQ("d", secondGeneration.GetValue<std::string>("d"));
 
     // New Gen
     thirdGeneration = secondGeneration.WithValue("e", std::string("e"));
@@ -234,13 +234,13 @@ TEST(Context, HeapLinkIntegrity)
   // Went out of scope, root and secondGeneration are destroyed. but should remain in heap for the
   // third-generation since the previous geneations are still alive inside his heart <3.
   EXPECT_TRUE(thirdGeneration.HasKey("a"));
-  EXPECT_EQ("a", thirdGeneration.Get<std::string>("a"));
+  EXPECT_EQ("a", thirdGeneration.GetValue<std::string>("a"));
   EXPECT_TRUE(thirdGeneration.HasKey("b"));
-  EXPECT_EQ("b", thirdGeneration.Get<std::string>("b"));
+  EXPECT_EQ("b", thirdGeneration.GetValue<std::string>("b"));
   EXPECT_TRUE(thirdGeneration.HasKey("c"));
-  EXPECT_EQ("c", thirdGeneration.Get<std::string>("c"));
+  EXPECT_EQ("c", thirdGeneration.GetValue<std::string>("c"));
   EXPECT_TRUE(thirdGeneration.HasKey("d"));
-  EXPECT_EQ("d", thirdGeneration.Get<std::string>("d"));
+  EXPECT_EQ("d", thirdGeneration.GetValue<std::string>("d"));
   EXPECT_TRUE(thirdGeneration.HasKey("e"));
-  EXPECT_EQ("e", thirdGeneration.Get<std::string>("e"));
+  EXPECT_EQ("e", thirdGeneration.GetValue<std::string>("e"));
 }